+++ src/actor.js
@@ -383,7 +383,7 @@ export const call = async (actId, input, options = {}) => {
 };
 
 /**
- * Runs an actor actor on the Apify platform using the current user account (determined by the `APIFY_TOKEN` environment variable),
+ * Runs an actor on the Apify platform using the current user account (determined by the `APIFY_TOKEN` environment variable),
  * waits for the actor to finish and fetches its output.
  *
  * By passing the `waitSecs` option you can reduce the maximum amount of time to wait for the run to finish.